swap modules context and typing context
